move schema version type to its own file
